Extend the FFM layout audit to the Unix platforms, and fix what it found (#3701)
* Extend the FFM audit to the Unix platforms, and fix three parse bugs it exposed

Seeds includes/ and structs/ for AIX, DragonFly BSD, FreeBSD, OpenBSD and Solaris. The
Unix variants share oshi.ffm.platform.unix, whose two layouts are platform-dependent, so
--sources is now repeatable: a platform's own package plus the shared one.

A header is optional when written with a leading `?`. cups is a package rather than a base
header on the BSDs, and a hard #include of a missing one takes the whole compile down
instead of reporting the two cups layouts as unresolvable, which is the honest answer.

Three harness bugs, all found by running this against real headers:

The inline-nested-layout branch of _size_of matched its closing paren with a greedy `.*`,
which ran past it into the trailing .withName(...) and pulled that text into the last
element. That element then sized as whatever token happened to start it -- by luck the
right answer for the Windows PDH nesting, and None for macOS utmpx, which silently
demoted the layout to unresolvable. Match the paren instead.

The layout regex required MemoryLayout and .structLayout to be adjacent, so spotless
wrapping a declaration between them dropped that layout from the audit entirely, with
nothing but a lower count to notice it by. Allow whitespace, and add an independent
guard: a StructLayout declared from a layout literal that was not extracted is now a
hard error rather than silently missing coverage.

Read sources as UTF-8 explicitly rather than at the platform default.

* Fix the two native-binding bugs the Unix audit found

Both were verified against the real headers on the GCC compile farm rather than inferred.

FreeBSD's `struct utmpx` is 280 bytes, not 276: ut_tv aligns it to 8 and structLayout
never infers tail padding. Every field offset was already right, so this was latent, in
the same way the Windows tail-padding errors were. Its JNA twin nests a Timeval and so
gets the size from JNA's own alignment.

AIX `thread_self` returns tid_t, which follows long -- measured at 8 bytes under -maix64
and 4 under -maix32. OSHI runs 64-bit there, and both twins declared it 4, reading only
the low half of the returned register. Declare it 64-bit and narrow at the API boundary,
since OSThread.getThreadId() is int and cannot widen.

Nest ut_tv in the macOS, Linux, Solaris and FreeBSD utmpx layouts rather than flattening
it to ut_tv_sec/ut_tv_usec, and give Linux's two spare arrays the names glibc declares
(__f_spare, __glibc_reserved). A member the header does not declare cannot be checked, so
those six offsets were being skipped on every platform -- on the very struct that turned
out to have a real bug on FreeBSD. Offsets are unchanged; macOS and Linux now check 164
and 43 field offsets with none unverifiable, against 163 and 40 with six skipped.

* Audit the BSD and Solaris layouts in CI

FreeBSD, OpenBSD and Solaris each declare layouts of their own, and they are the ones
nobody can eyeball. The VMs need no JDK -- the audit is python plus the base system's
compiler and headers -- so these are much cheaper than the test jobs on the same images.

AIX and DragonFly BSD are left to the compile farm. Neither declares a layout of its own,
so both only exercise the shared unix package that the other three already cover.

* Do not name statvfs's reserved tail, whose shape is a glibc version

Renaming the field to what glibc calls it made it checkable, and the check promptly
disagreed with itself across distributions: up to glibc 2.36 the tail is __f_spare[6],
and from 2.37 an f_type was carved out of its front leaving __f_spare[5]. Debian 12
passed and Ubuntu 24.04 failed on the same tree.

The total is 112 bytes either way and OSHI reads none of it, so naming it asserts a glibc
version OSHI has no need of and no single name is right. Model it as unnamed padding and
let the size assert carry the check. Verified clean against both glibc 2.36 and 2.39.

* Map AIX thread_self to NativeLong in JNA, not a fixed 64-bit long

tid_t follows long -- measured at 4 bytes compiling with -maix32 and 8 with -maix64 --
which is exactly what NativeLong exists for. Declaring `long` hard-codes int64 and would
read 8 bytes for a 4-byte return under a 32-bit JVM. FreeBSD's thr_self is the existing
precedent in this repository, and takes a NativeLongByReference for the same reason.

The FFM twin keeps JAVA_LONG: it has no NativeLong, and Java 25 on AIX is 64-bit only, the
same assumption SIZE_T in that file already states. Its comment now says so, and says why
the two differ. (Linker.canonicalLayouts().get("long") is the FFM analogue if that ever
stops holding, at the cost of a carrier that varies by platform.)

Verified on the AIX box: getThreadId() returns 107807033 through the new mapping, which is
also what the old int declaration read -- the tid fits in 31 bits, so this stays latent.

+++ oshi-core-ffm/src/main/java/oshi/software/os/unix/aix/AixOperatingSystemFFM.java
@@ -62,7 +62,8 @@ public final class AixOperatingSystemFFM extends AixOperatingSystem {
     @Override
     public int getThreadId() {
         try {
-            return AixLibcFunctions.thread_self();
+            // OSThread.getThreadId() is int, so narrow at the API boundary
+            return (int) AixLibcFunctions.thread_self();
         } catch (Throwable _) {
             return 0;
         }
